ELEMENTS-1142: moved widgets into individual folders
